Refactoring and move plot
