+++ pertpy/metadata/_look_up.py
@@ -24,10 +24,10 @@ class LookUp:
     ):
         """
         Args:
-            type: metadata type for annotation, cell_line, compound or moa. Defaults to cell_line.
-            transfer_metadata: dataframes used to generate Lookup object. This is currently set to None for CompoundMetaData which does not require any dataframes for transfer. Defaults to None. To ensure efficient transfer of
-                metadata during initialization, LookUp object should always be generated by the corresponding MetaData
-                class. Also, different MetaData classes have different required metadata to transfer.
+            type: Metadata type for annotation. One of 'cell_line', 'compound' or 'moa'. Defaults to cell_line.
+            transfer_metadata: DataFrames used to generate Lookup object.
+                               This is currently set to None for CompoundMetaData which does not require any dataframes for transfer.
+                               Defaults to 'cell_line'.
         """
         if type == "cell_line":
             self.type = type
@@ -245,7 +245,6 @@ class LookUp:
                 "stripped_cell_line_name". Defaults to "DepMap_ID".
             query_id_list: Unique cell line identifiers to test the number of matched ids present in the
                 metadata. If set to None, the query of metadata identifiers will be disabled. Defaults to None.
-
         """
         if self.type != "cell_line":
             raise ValueError("This is not a LookUp object specifically for CellLineMetaData!")
@@ -336,12 +335,20 @@ class LookUp:
         """A brief summary of drug response data.
 
         Args:
-            gdsc_dataset: The GDSC dataset, 1 or 2. Defaults to 1. The GDSC1 dataset updates previous releases with additional drug screening data from the Wellcome Sanger Institute and Massachusetts General Hospital. It covers 970 Cell lines and 403 Compounds with 333292 IC50s. GDSC2 is new and has 243,466 IC50 results from the latest screening at the Wellcome Sanger Institute using improved experimental procedures.
-            reference_id: The type of cell line identifier in the meta data, cell_line_name, sanger_model_id or cosmic_id. Defaults to "cell_line_name".
-            query_id_list: Unique cell line identifiers to test the number of matched ids present in the metadata. If set to None, the query of metadata identifiers will be disabled. Defaults to None.
-            reference_perturbation: The perturbation information in the meta data, drug_name or drug_id. Defaults to "drug_name".
-            query_perturbation_list: Unique perturbation types to test the number of matched ones present in the metadata. If set to None, the query of perturbation types will be disabled. Defaults to None.
-
+            gdsc_dataset: The GDSC dataset, 1 or 2. Defaults to 1.
+                          The GDSC1 dataset updates previous releases with additional drug screening data from the Wellcome Sanger Institute and Massachusetts General Hospital.
+                          It covers 970 Cell lines and 403 Compounds with 333292 IC50s.
+                          GDSC2 is new and has 243,466 IC50 results from the latest screening at the Wellcome Sanger Institute using improved experimental procedures.
+            reference_id: The type of cell line identifier in the meta data, cell_line_name, sanger_model_id or cosmic_id.
+                          Defaults to 'cell_line_name'.
+            query_id_list: Unique cell line identifiers to test the number of matched ids present in the metadata.
+                           If set to None, the query of metadata identifiers will be disabled.
+                           Defaults to None.
+            reference_perturbation: The perturbation information in the meta data, drug_name or drug_id.
+                                    Defaults to 'drug_name'.
+            query_perturbation_list: Unique perturbation types to test the number of matched ones present in the metadata.
+                                     If set to None, the query of perturbation types will be disabled.
+                                     Defaults to None.
         """
         if self.type != "cell_line":
             raise ValueError("This is not a LookUp object specific for CellLineMetaData!")
@@ -378,8 +385,10 @@ class LookUp:
         """A brief summary of gene annotation metadata
 
         Args:
-            reference_id: The type of gene identifier in the meta data, gene_id, ensembl_gene_id, hgnc_id, hgnc_symbol. Defaults to "ensembl_gene_id".
-            query_id_list: Unique gene identifiers to test the number of matched ids present in the metadata. Defaults to None.
+            reference_id: The type of gene identifier in the meta data, gene_id, ensembl_gene_id, hgnc_id, hgnc_symbol.
+                          Defaults to "ensembl_gene_id".
+            query_id_list: Unique gene identifiers to test the number of matched ids present in the metadata.
+                           Defaults to None.
         """
         if self.type != "cell_line":
             raise ValueError("This is not a LookUp object specific for CellLineMetaData!")
@@ -415,8 +424,12 @@ class LookUp:
         """A brief summary of MoA annotation.
 
         Args:
-            query_id_list: Unique perturbagens to test the number of matched ones present in the metadata. If set to None, the query of metadata perturbagens will be disabled. Defaults to None.
-            target_list: Unique molecular targets to test the number of matched ones present in the metadata. If set to None, the comparison of molecular targets in the query of metadata perturbagens will be disabled. Defaults to None.
+            query_id_list: Unique perturbagens to test the number of matched ones present in the metadata.
+                           If set to None, the query of metadata perturbagens will be disabled.
+                           Defaults to None.
+            target_list: Unique molecular targets to test the number of matched ones present in the metadata.
+                         If set to None, the comparison of molecular targets in the query of metadata perturbagens will be disabled.
+                         Defaults to None.
         """
         if self.type != "moa":
             raise ValueError("This is not a LookUp object specific for MoaMetaData!")
@@ -443,8 +456,10 @@ def available_compounds(
     """A brief summary of compound annotation.
 
     Args:
-        query_id_list: Unique compounds to test the number of matched ones present in the metadata. If set to None, query of compound identifiers will be disabled. Defaults to None.
-        query_id_type: The type of compound identifiers, name or cid. Defaults to "name".
+        query_id_list: Unique compounds to test the number of matched ones present in the metadata.
+                       If set to None, query of compound identifiers will be disabled.
+                       Defaults to None.
+        query_id_type: The type of compound identifiers, name or cid. Defaults to 'name'.
     """
     if self.type != "compound":
         raise ValueError("This is not a LookUp object specific for CompoundData!")
