+++ pertpy/plot/__init__.py
@@ -1,6 +1,5 @@
 from pertpy.plot._augur import AugurpyPlot as ag
 from pertpy.plot._coda import CodaPlot as coda
 from pertpy.plot._guide_rna import GuideRnaPlot as guide
-from pertpy.plot._metadata import MetaDataPlot as md
 from pertpy.plot._milopy import MilopyPlot as milo
 from pertpy.plot._mixscape import MixscapePlot as ms
